Rename minApiLevel method
This method was originally created to fetch the API level for minimum API usage.  The logic of the method itself doesn't really pertain to minimum or deprecation - it's just finding a Doc Tag from the data elements.  Thus, this method should be renamed apiLevel() or something similar, with the "min" part removed to avoid confusion.

Bug: 129726096