Fix Gradle plugin application
Default dependency to dokka-base in dokkaPluginConfiguration was changed to a regular dependency. Applying client-side plugins overrode the default dokka-base dependency thus changing the version to a (possibly) lower one. This could lead to unpredictable Dokka crashes due to the inconsistency between the Gradle runner and dokka-base versions. Dependencies can still be manually removed from the configuration in specific projects, if needed (not likely)