Fix MathJax rendering bug (#2175)
Fix a bug where using the `@usesMathJax` custom tag would cause the
documentation to fail to render. I fixed this by making all custom tags
render to the screen, which were being ignored.

Along with this, I updated the existing test case to check for the
presence of content in the output.

Fixes #2175