Avoid snapshotting sourceLink.localDirectory input
A typical use case in a project with complex source root setup is to set sourceLink.localDirectory to the root of the project checkout dir and remoteUrl to the root of the project VCS url respectively.
However, since this property was marked as InputDirectory, in that case Gradle snapshotted hashes of all files in this entire project directory before running the Dokka task (including for example .git and .gradle directories). This could be very slow in big projects and also result in various Gradle warnings.

Now localDirectory is not an input property, but its relativized path is stored in another read-only input string property. Therefore, it invalidates task outputs when its value changes, but not when any content inside it.