Introduce InternalDokkaApi annotation
Rationale:

dokka-core has a long history of bloating its API shape with utilities that were never intended to be public and that may expose unwanted implementation details, as well as unwanted compatibility burden.

Eventually, we would like to get rid of them (i.e. by making them internal), but first it would be nice to provide users with an explicit message about it