Add `dokkaPlugin` configuration to example code
In the process of trying to apply a `dokkaPlugin`, it was in example code to include this line but not mentioned in the documentation. This is a one line change to add that line to the Kotlin script example. I'm not certain if this also applies to the Maven and Groovy implementations, but I wanted to at least start the conversation