Fix cross module resolution local links (#3425)
Fixes #2272. There is a situation where 2 (or more) gradle modules have the same package and dokka doesn’t know where to link.
E.g.
The class com.example.classA is in moduleA
and com.example.classB is in moduleB
moduleC depends on moduleA and moduleB.
In moduleC: there is a link [com.example.classB]. Having only package-list, dokka doesn’t know where to link, since the package com.example is in both modules.

This is short-term and temporary solution of #3368 to cover a case of links between only local modules in a project.
It is based on #2901 from @EddieRingle. It generates an external link depending on a check if the local link exists. The check works only when the number of possible resolved links are more than one.



* all-modules-page: Validate cross-module links during resolution

The existing behavior was just taking the first module that includes the
package name of a given reference, so this resulted in broken links when
a package is split across multiple modules.

This change checks to ensure the resolved file exists before choosing it
as the target for the resolved link.

Signed-off-by: Eddie Ringle <eddie@ringle.io>

* Check links only if resolved links has more than one link

* Add integration test

* Refactor

* Add comment

---------

Signed-off-by: Eddie Ringle <eddie@ringle.io>
Co-authored-by: Eddie Ringle <eddie@ringle.io>