add 'enabled' flag for Build Scans
GitHub can't have default environment variables with dots `.` in them, so defining the
build scan server URL in all actions would have to be manually done for each GitHub
action. That's annoying and is work to maintain.

Instead, this PR adds a 'build.scan.enabled' flag that defaults to 'true' on CI.