Update Maven Binary install task to be cachable
Currently the `installMavenBinary` task does not normalize the task inputs. This results in the task re-running unnecessarily.

To improve build performance, use classpath-normalization on the Maven files before copying them to a local directory.

Additionally, update the Configurations to follow Gradle conventions (create a 'task bucket' Configuration for declaring dependencies, and a 'resolver' to fetch the dependencies.)