Reduce log level of test Gradle projects (#3717)
* Reduce log level of test Gradle projects

### Motivation

Too many logs when using Gradle TestKit causes the tests to run slower. This is mainly because Gradle TestKit stores logs in-memory (see gradle/gradle#23965).

Reducing the amount of logging would be beneficial for two reasons:

- The tests can run faster.
- The logs are not cluttered with unimportant messages from other Gradle tasks. We can focus on the Dokka Generator logs.

However, reducing the log level to 'lifecycle' would mean that the Dokka Generator logs would not be reported. To work around this, I added an internal flag that can override the default log level of Dokka Generator.

The result is that the tests run faster, and test results are easier to read and verify.

### Summary of changes

- Reduce the log-level of Gradle tests to 'lifecycle' by default (this can be overridden if necessary).
- Introduce an internal flag, `dokka.internal.dokkaGenerator.logLevelOverride`, in AbstractDokkaTask that can re-direct the logs of Dokka Generator to a specific level, independent from the Gradle log level.
- Correctly enable `--stacktrace` on all tests, and enable it by default for all tests.

^OSIP-355

* rename property

* ignore invalid log level values