Prettify file trees in docs (#3754)
Minor docs formatting update, so that the file trees look better. E.g. use `└` for the last file in a list instead of `├`.

I know these docs are probably going to be outdated soon, but the file trees not being pretty bugs me a little.

I used https://tree.nathanfriend.io/ to update the trees.