DGP: Rename conflicting 'maybeCreate' helper function.
DGPv1 and v2 both have a similar utility function: `fun <T> NamedDomainObjectContainer<T>.maybeCreate`. However, their behaviour is slightly different.

v1: maybe create the item, OR if it exists, configure it.
v2: maybe create the item, AND exist the new/existing item.

The only usage of v1's version was in the `gradleConfiguration.kt`, so this PR moves it and renames it to `findOrCreate()`, to avoid confusion.