Rename `dokkaModuleDirectory`, and mark as internal
Follow up for #3876

When Dokka produces 'partial' Modules it stores them in a project-local directory.

How Dokka processes Modules is an implementation detail, so exposing the directory is not useful for users, and could be confusing.
It is useful to have a single property for DGP to use though, so it should be kept.

As such, mark it as `internal` to hide it from the DSL.