Remove V2EnabledWithHelpers warning & info message
When users update to 2.1.0, V2EnabledWithHelpers is enabled automatically.

The info message was intended to nude people to migrate, but in 2.1.0 there's no need to nudge.
The other intention was to add links to the migration guide, but the deprecation message contains a link to the guide.

Without removing the message, we will require _all_ users to suppress the warning.
This is an annoyance, and we should just remove the message.