DPG: Only apply 'Adapter' plugins once, and group PluginIds
The AndroidAdapter, JavaAdapter, and KotlinAdapters apply their configuration multiple times.

Currently, this is acceptable because a project cannot have multiple Android/Kotlin/Java plugins applied.
But the situation gets more complicated with AGP9 and kotlin-built-in.
The kotlin-built-in plugin uses an implementation of `KotlinBasePlugin`,
so the Android and Kotlin adapters could get applied multiple times, and some of their operations (registering DokkaSourceSets) aren't idempotent.

I also updated the JavaAdapter to use the same approach for consistency (it's not strictly necessary).

---

Also, group the plugin IDs into Sets, so they can be accessed at once without needing to remember which IDs should be accessed.