DPG IT: Test `com.android.kotlin.multiplatform.library` with Kotlin Multiplatform (#4359)
Add a simple test for a Android KMP library project.

It's just a simple test without extended KMP targets to keep it simpler. There are other projects for testing KMP targets. The main point of this test is to verify DGP can extract info (especially `android.jar`) from the Android KMP library plugin.

There's a workaround for an AGP8 bug that will be fixed in AGP9.

Fix #4211