Rework the Gradle attributes to make them integrate better with the ecosystem (#4334)
Unfortunately, Gradle's matching algorithm prioritises the number of attributes matched over the actual attribute values. (Note that if a candidate does not have a requested attribute this still counts as a match!) This behaviour makes it impossible to strictly match Configurations, and leads to situations like this where two 'ecosystems' step on each other's toes.

The solution is in two parts:

1. Reduce the number of attributes DGP uses.

   This prevents the matching algorithm from over-matching based on attributes alone. The Bundling attribute is removed, since this is not important for DGP classpaths.
4. (As suggested by @martinbonnin) DGP configurations use a custom value for the `Usage` attribute, and use a one-way `AttributeCompatibilityRule` so DGP can still resolve dependencies.
   (A one-way rule should minimise ugly issues [where rules can leak](https://github.com/gradle/gradle/issues/30921).)

See the KDoc of `DokkaJavaRuntimeUsageCompatibilityRule` for details.


---------

Co-authored-by: Martin Bonnin <martin@mbonnin.net>