Only apply 'Adapter' plugins once, and group PluginIds (#4362)
The AndroidAdapter, JavaAdapter, and KotlinAdapters apply their configuration multiple times.

Currently, this is acceptable because a project cannot have multiple Android/Kotlin/Java plugins applied.
But the situation gets more complicated with AGP9 and kotlin-built-in.
The kotlin-built-in plugin uses an implementation of `KotlinBasePlugin`,
so the Android and Kotlin adapters could get applied multiple times, and some of their operations (registering DokkaSourceSets) aren't idempotent.

I also updated the JavaAdapter to use the same approach for consistency (it's not strictly necessary).

---

add 'test' and 'dynamic-feature' Android plugins

(Adding them is not strictly related to this PR. They're added as part of the larger initiative of AGP9 support. We're not adding tests for them because I can't find any up-to-date examples, but they should just work.

---

AndroidAdapter: remove unnecessary `project.plugins.withType<DokkaBasePlugin>().all { }`§