KGP IT: Support testing AGP with Kotlin-built-in
Add support for testing AGP projects that use kotlin-built-in.

I converted the existing `requiredAgpMajorVersions` check into a unit test.
This will help verify that TestedVersionsSource doesn't accidentally filter out any of the tested versions.
It could be a bit of a pain to maintain? But this test is quick and easy, and if it prevents a bug where a version is accidentally excluded, it'd be worth it.
(We could use test coverage support on TeamCity? But I'm not sure how to add that, and it could be fragile or easy to ignore?)

Currently there are no test-projects for kotlin-built-in. These will be added in #4295.